Fixed #152. TODO: Improve logging
